MQTT outbound connector doesn't use connection properties.
