+++ mqtt-transport/src/main/resources/com/esri/geoevent/transport/mqtt-transport_en_US.properties
@@ -36,11 +36,8 @@ TRANSPORT_COMMON_QOS_2_LBL=2 (exactly once)
 # Log Messages
 CONNECTION_TEST_FAILED=Failed to connect to {0}. Error: {1}.
 CONNECTION_LOST=Connection lost: {0}
-UNEXPECTED_ERROR=Unexpected error during MQTT receiver initialization.
-UNEXPECTED_ERROR_STARTING=Unexpected error when starting the Transport.
+UNEXPECTED_ERROR=Unexpected Error: {0}.
 BUFFER_OVERFLOW_ERROR=Buffer overflow.  Flushing the buffer and continuing.
-UNEXPECTED_ERROR2=Unexpected error, stopping the Transport.
 UNABLE_TO_CLOSE=Unable to close MQTT client.
-
 INIT_ERROR=Unable to initialize the {0} transport.
 ERROR_PUBLISHING=ERROR occurred while publishing message: {0}
